chore: 🤖 Update makefile
✅ Closes: git status
